Add continuous integration data design
This is the first in a long line of commits that will help keep my world
sane. The idea here is that the way that modern CI systems operate is
terrible. I am only familiar with Jenkins and Hudson, but the overall
structure is more less forgiving when it comes to validating software
across multiple platforms. Some developers give in to the mantra of "well
it works for me, and compiles everywhere else, that should be
sufficient." This is terrible. If you are writing for multiple
environments, you need to build and test in them.

This is the point of this commit. More will follow until the entire
application is rewritten from the ground up and eating its own dog-food.